backend/requirements.txt
pyramid>=2.0
waitress>=3.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
transformers>=4.35
torch>=2.0
google-generativeai>=0.3
python-dotenv>=1.0
requests>=2.31
